requirements.txt
-e .
aws-cdk.core==1.70.0
aws-cdk.aws-s3==1.70.0

